refactor queries by breaking them up into seperate files
