+++ src/apps/experimental/components/library/GenresItemsContainer.tsx
@@ -1,11 +1,11 @@
 import type { BaseItemKind } from '@jellyfin/sdk/lib/generated-client/models/base-item-kind';
 import type { CollectionType } from '@jellyfin/sdk/lib/generated-client/models/collection-type';
 import React, { FC } from 'react';
-import { useGetGenres } from 'hooks/useFetchItems';
 import globalize from 'scripts/globalize';
 import Loading from 'components/loading/LoadingComponent';
 import GenresSectionContainer from './GenresSectionContainer';
 import type { ParentId } from 'types/library';
+import { useGetGenres } from 'hooks/useQueries/useGetGenres';
 
 interface GenresItemsContainerProps {
     parentId: ParentId;
