+++ src/apps/experimental/components/library/ItemsView.tsx
@@ -6,7 +6,6 @@ import React, { type FC, useCallback } from 'react';
 import Box from '@mui/material/Box';
 import classNames from 'classnames';
 import { useLocalStorage } from 'hooks/useLocalStorage';
-import { useGetItemsViewByType } from 'hooks/useFetchItems';
 import { getDefaultLibraryViewSettings, getSettingsKey } from 'utils/items';
 import { CardShape } from 'utils/card';
 import Loading from 'components/loading/LoadingComponent';
@@ -29,6 +28,7 @@ import { type LibraryViewSettings, type ParentId, ViewMode } from 'types/library
 import type { CardOptions } from 'types/cardOptions';
 import type { ListOptions } from 'types/listOptions';
 import { useItem } from 'hooks/useItem';
+import { useGetItemsViewByType } from 'hooks/useQueries/useGetItemsViewByType';
 
 interface ItemsViewProps {
     viewType: LibraryTab;
