+++ src/apps/experimental/components/library/UpcomingView.tsx
@@ -1,11 +1,11 @@
 import React, { type FC } from 'react';
 import Box from '@mui/material/Box';
-import { useGetGroupsUpcomingEpisodes } from 'hooks/useFetchItems';
 import Loading from 'components/loading/LoadingComponent';
 import globalize from 'scripts/globalize';
 import SectionContainer from './SectionContainer';
 import { CardShape } from 'utils/card';
 import type { LibraryViewProps } from 'types/library';
+import { useGetGroupsUpcomingEpisodes } from 'hooks/useQueries/useGetGroupsUpcomingEpisodes';
 
 const UpcomingView: FC<LibraryViewProps> = ({ parentId }) => {
     const { isLoading, data: groupsUpcomingEpisodes } = useGetGroupsUpcomingEpisodes(parentId);
