+++ src/hooks/useQueries/useGetItems.tsx
@@ -1,0 +1,45 @@
+import type { AxiosRequestConfig } from 'axios';
+import {
+    ItemsApiGetItemsRequest,
+    ItemSortBy
+} from '@jellyfin/sdk/lib/generated-client';
+import { getItemsApi } from '@jellyfin/sdk/lib/utils/api/items-api';
+import { useQuery } from '@tanstack/react-query';
+import { JellyfinApiContext, useApi } from 'hooks/useApi';
+
+const fetchGetItems = async (
+    currentApi: JellyfinApiContext,
+    parametersOptions: ItemsApiGetItemsRequest,
+    options?: AxiosRequestConfig
+) => {
+    const { api, user } = currentApi;
+    if (api && user?.Id) {
+        const response = await getItemsApi(api).getItems(
+            {
+                userId: user.Id,
+                ...parametersOptions
+            },
+            {
+                signal: options?.signal
+            }
+        );
+        return response.data;
+    }
+};
+
+export const useGetItems = (parametersOptions: ItemsApiGetItemsRequest) => {
+    const currentApi = useApi();
+    return useQuery({
+        queryKey: [
+            'Items',
+            {
+                ...parametersOptions
+            }
+        ],
+        queryFn: ({ signal }) =>
+            fetchGetItems(currentApi, parametersOptions, { signal }),
+        gcTime: parametersOptions.sortBy?.includes(ItemSortBy.Random) ?
+            0 :
+            undefined
+    });
+};
