+++ src/hooks/useQueries/useGetQueryFiltersLegacy.tsx
@@ -1,0 +1,44 @@
+import type { AxiosRequestConfig } from 'axios';
+import { BaseItemKind } from '@jellyfin/sdk/lib/generated-client';
+import { getFilterApi } from '@jellyfin/sdk/lib/utils/api/filter-api';
+import { useQuery } from '@tanstack/react-query';
+import { JellyfinApiContext, useApi } from 'hooks/useApi';
+import { ParentId } from 'types/library';
+
+const fetchGetQueryFiltersLegacy = async (
+    currentApi: JellyfinApiContext,
+    parentId: ParentId,
+    itemType: BaseItemKind[],
+    options?: AxiosRequestConfig
+) => {
+    const { api, user } = currentApi;
+    if (api && user?.Id) {
+        const response = await getFilterApi(api).getQueryFiltersLegacy(
+            {
+                userId: user.Id,
+                parentId: parentId ?? undefined,
+                includeItemTypes: itemType
+            },
+            {
+                signal: options?.signal
+            }
+        );
+        return response.data;
+    }
+};
+
+export const useGetQueryFiltersLegacy = (
+    parentId: ParentId,
+    itemType: BaseItemKind[]
+) => {
+    const currentApi = useApi();
+    const isLivetv = parentId === 'livetv';
+    return useQuery({
+        queryKey: ['QueryFiltersLegacy', parentId, itemType],
+        queryFn: ({ signal }) =>
+            fetchGetQueryFiltersLegacy(currentApi, parentId, itemType, {
+                signal
+            }),
+        enabled: !!parentId && !isLivetv
+    });
+};
