+++ src/hooks/useQueries/useGetStudios.tsx
@@ -1,0 +1,49 @@
+import type { AxiosRequestConfig } from 'axios';
+import {
+    BaseItemKind,
+    ItemFields,
+    ImageType
+} from '@jellyfin/sdk/lib/generated-client';
+import { getStudiosApi } from '@jellyfin/sdk/lib/utils/api/studios-api';
+import { useQuery } from '@tanstack/react-query';
+import { JellyfinApiContext, useApi } from 'hooks/useApi';
+import { ParentId } from 'types/library';
+
+const fetchGetStudios = async (
+    currentApi: JellyfinApiContext,
+    parentId: ParentId,
+    itemType: BaseItemKind[],
+    options?: AxiosRequestConfig
+) => {
+    const { api, user } = currentApi;
+    if (api && user?.Id) {
+        const response = await getStudiosApi(api).getStudios(
+            {
+                userId: user.Id,
+                includeItemTypes: itemType,
+                fields: [
+                    ItemFields.DateCreated,
+                    ItemFields.PrimaryImageAspectRatio
+                ],
+                enableImageTypes: [ImageType.Thumb],
+                parentId: parentId ?? undefined,
+                enableTotalRecordCount: false
+            },
+            {
+                signal: options?.signal
+            }
+        );
+        return response.data.Items;
+    }
+};
+
+export const useGetStudios = (parentId: ParentId, itemType: BaseItemKind[]) => {
+    const currentApi = useApi();
+    const isLivetv = parentId === 'livetv';
+    return useQuery({
+        queryKey: ['Studios', parentId, itemType],
+        queryFn: ({ signal }) =>
+            fetchGetStudios(currentApi, parentId, itemType, { signal }),
+        enabled: !!parentId && !isLivetv
+    });
+};
